Cleanup and API revisions

+++ carbon/include/flipper/atmegau2/megausb.h
@@ -60,13 +60,12 @@ extern const struct descriptor {
 	uint8_t length;
 } PROGMEM descriptors[NUM_DESC_LIST];
 
-int8_t megausb_interrupt_receive(void *dst, uint32_t length);
-int8_t megausb_interrupt_transmit(void *src, uint32_t length);
+int megausb_interrupt_receive(void *dst, uint32_t length);
+int megausb_interrupt_transmit(void *src, uint32_t length);
 
-int8_t megausb_bulk_receive(void *dst, uint32_t length);
-int8_t megausb_bulk_transmit(void *src, uint32_t length);
+int megausb_bulk_receive(void *dst, uint32_t length);
+int megausb_bulk_transmit(void *src, uint32_t length);
 
 int usb_debug_putchar(uint8_t c);
 
 #endif
-
